Change integer to bigInteger at no_telp

+++ database/migrations/2024_06_13_125039_create_customers_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -16,7 +15,7 @@ return new class extends Migration
         Schema::create('customers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('no_telp');
+            $table->bigInteger('no_telp');
             $table->text('address');
             $table->timestamps();
         });
